tux3: 64bit inum cleanup
This introduce tux_set_inum() to set the inode number, and this hides
the detail of it.

So, we use tux_inode(inode)->inum for inode number, and in kernel,
inode->i_ino is just used as hash value (low 32bit of inum).

Then, this uses iget5_locked to check ->inum, and adds tux3_getattr()
to export inode number to userland. With LFS (Large File Support),
stat64->ino and dirent64->ino are 64bit in userland.

With this, basically tux3 can use 64bit inum cleanly. The issue is
users in kernel, and non-LFS users. E.g. audit (CONFIG_AUDIT) uses
inode->i_ino directly to check if it is interesting inode.  So, later,
we may need to add the limit of inode number for users of those, or
fix those instead.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>
Signed-off-by: Daniel Phillips <daniel@tux3.org>