tux3: Use new blockdirty() strategy
With new blockdirty strategy, we don't need to care about
block-fork. So, this does list_splice_init() to get stable list for
flushing delta, and flush the buffers on list.

And when clean the buffer, call evict_buffer() early if the buffer is forked.
Because, forked buffer is unneeded anymore.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>
Signed-off-by: Daniel Phillips <daniel@tux3.org>