tux3: Fix the cursor position handling of insert_leaf()
Current code updates the childblock/childkey to the splited new block.
And it updates cursor posision to childblock. It means it is assuming
the splited new block has inserting leaf.

But, it is not true. So, this patch fixes the cursor position by
changing the position only if splited new block has the target leaf.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>
Signed-off-by: Daniel Phillips <daniel@tux3.org>