tux3: Move flushing codes to writeback.c
This moves and tweaks the flushing codes like tuxsync() and
sync_super() to writeback.c.

Then, this adds the dummy codes of those functions to
writeback.c. With dummy, source code dependency would be solved. If
binary is not including the real functions, writeback.c will provide
dummy code instead.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>
Signed-off-by: Daniel Phillips <daniel@tux3.org>