tux3: Take the inode refcnt for writeback
This takes the refcnt of inode when inode became dirty. With this,
inodes can be iput() before flushing the inode, and also don't need to
flush inode every operations. sync_super() flushes the all dirty inodes.

BTW, this has difference with kernel. Because we don't cache the
inodes, so we take refcnt when inode became dirty.  So, inodes never
be freed while it's dirty. (on kernel, it will flush the inode when
inode was expired)

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>
Signed-off-by: Daniel Phillips <daniel@tux3.org>