tux3: Separate dir inode update from dir entry operations (delete path)
The directory format is also used for xattr. And for now, the
functions are shared from both. So, if the operation is for xattr, it
will update the inode (e.g. timestamp) too. But, this update is
unneeded, and it will become the cause of update of atable inode on
ileaf.

So, this separates those from *_entry() functions. Fortunately, it
seems to be able to separate cleanly at least for now. With this
series, the entry operation is called *_entry(), and adds the
*_dirent() functions for directory (to update the inode too).

This patch is for delete path.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>
Signed-off-by: Daniel Phillips <daniel@tux3.org>