tux3: Prepare to use LOG_ROLLUP log
We have to know whether there is LOG_ROLLUP or not, and position of it.
Also, we are already makeing block address array for logblock, but it
is doing for each the replay stage.

So, this prepares both of block address and LOG_ROLLUP check in early
as replay_prepare().

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>