tux3: Fix inode->i_size truncation on tux3_rmdir()
On tux3_rmdir(), it set 0 to inode->i_size. But, it confuses
tuxtruncate(), because it checks inode->i_size to know whether
btree_chop() is needed or not.

That dependency would not be good though, for now, fix it by changing
tux3_rmdir(). Also, there is no good reason to set 0.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>