tux3: Remove i_blocks partial implements
FIXME: We need to implement ->i_blocks somehow?

If we want to add i_blocks account, we have to check existent extent
for dirty buffer.  And only if there is no existent extent, we add to
->i_blocks.

Yes, ->i_blocks must be including delayed allocation buffers as
allocated, because some apps (e.g. tar) think it is empty file if
i_blocks == 0.

But, it is purely unnecessary overhead.

And it is just used as hint (if i_blocks < align(i_size, blocksize),
this file would be sparse) to optimize file copy on userland. For
optimization of userland, we should implement SEEK_HOLE/DATA instead.

So, for now, we don't use i_blocks at all, instead just set st_blocks
at stat(2).

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>