tux3: tux3graph, Cleanup link of btree bnode/leaf
Now, we are using self descriptive name (e.g. bnode_30, dleaf_20) to
draw link. However, to do it, we have to know the detail of buffer is
what kind of data type.

Instead of it, this just use simple name (e.g. volmap_30). This name
is not descriptive though, the code become simple and can make link
without buffer detail.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>