tux3: Fix clearing dirty flag on page for volmap
For the volmap, we can't clear the dirty flag on the page simply.

Because the page for volmap may have the dirty buffer for both of
"delta" and "rollup", and we may flush only dirty buffers for
"delta". So, if the page still has the dirty buffer, we should still
keep the page dirty for "rollup".

NOTE, this implement doesn't do test_and_set atomicity. So, instead,
we should serialize the backend itself (modification of volmap).

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>