tux3: Fix cursor path corruption after insert_leaf()
In insert_leaf(), we have to adjust cursor position after adding new
leaf. But there is the bug to adjusting cursor position.

If new child is in left node after split, we don't have to adjust
cursor position of parent. But if new child is right node after split,
we have to change cursor position of parent.

But current implement is not adjusting correctly.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>