tux3: Cleanup DATA_BTREE_BIT usage
The btree root is backend object, so it is only changed by
backend. And it is not good idea to change stabled ->present in
backend.

Instead of checking ->present, this just check whether btree itself
has data. If it has data, we save btree root, and set DATA_BTREE_BIT
in encode_attrs().

And also, in decode_attrs(), we don't set DATA_BTREE_BIT to ->preset.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>