tux3: Front/back separation, Introduce hole extents
To separate frontend and backend, we have to delay truncate (and
punch hole in future) operation.

The hole extents works as middle layer of page cache and dtree.

When read data, frontend checks page cache at first. Then, if there is
no page cache, it lookups dtree to get address of data.

To delay truncate of dtree, this adds the hole extents for truncated
region before lookup dtree, like delayed allocation. With this,
frontend doesn't lookup dtree if there is the hole extents.

And backend will apply the hole extents to dtree later, and do actual
truncation and freeing blocks.

This patch adds structure for hole extents, and slab initialization.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>