tux3: Replace mark_buffer_dirty() in block_truncate_block()
We have to manage dirty buffers, and it was done when marking buffer
dirty in tux3_mark_buffer_dirty().  But block_truncate_block() calls
mark_buffer_dirty() directly, so we have to replace it by
tux3_mark_buffer_dirty().

This fix it by copying block_truncate_block() to replace that.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>