tux3: Consolidate dirty inode, and fix
Consolidate dirty inode with userland and kernel.  Additioanlly, this
fixes the issue of I_DIRTY_PAGES.

->dirty_inode() is not called for I_DIRTY_PAGES. But we have to handle
that case too. Because we want to know I_DIRTY_PAGES, and list up
dirty inodes. So this patch adds wrapper for __mark_inode_dirty() as
__tux3_mark_inode_dirty(). (we use tux3_mark_buffer_dirty, so there is
no user of I_DIRTY_PAGES in vfs)

With this, we can add inode to dirty list for I_DIRTY_PAGES too.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>