tux3: Fix free_empty_btree() of redirect check
In free_empty_btree(), if buffer is not needed to redirect, we can
free the block immediately.  Current checks if buffer is dirty. But we
should check !cursor_need_redirect() instead.

So, this separates cursor_need_redirect() to leaf_need_redirect() and
bnode_need_redirect(). Then, use those in free_empty_btree().

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>