tux3: Fix canceling the dirty flag of buffer and page
For our writeback, when clear the dirty flag of buffer, we may have to
clear the dirty flag of page too.  Because clear_buffer_dirty() just
clear the dirty flag of buffer, and leaves the dirty flag of page as is.

Generic writeback traverses dirty pages in radix-tree, and write the
data pointed by dirty buffers. Also this writeback works as cleaner of
the left dirty flag of page. I.e. writeback didn't find any dirty
buffers on the page, just clear the dirty flag of page without I/O.

On tux3 writeback, we don't use radix-tree, instead, we use the list
of dirty buffers. This means, nobody clean the left dirty flag of
page.

Also we are using redirty_page_for_writepage() in ->writepage()
handler. This means, page reclaimer will see dirty page after called
->writepage(), and will not try to free buffer_head and will not clear
the dirty flag of page.

So, this patch fixes it by checking the dirty flag of page for each
clearing the dirty of buffer.

FIXME: this can be inefficient. We may want to check the dirty pages
with periodical batched check like writepages() does.

FIXME: cancel dirty is untested for mmap write.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>