tux3: Introduce TUX3_INODE_NO_FLUSH inode flag
For some types of inode, we don't add inode to dirty list. Because
some types of inode have own timing to flush (bitmap and volmap).

To do this, now, we are checking the inode number.

This introduces the inode flag called TUX3_INODE_NO_FLUSH, instead of
checking inode number. With this, we can easily check that type.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>