tux3: Fix balloc() for logblock itself in replay_log_stage2()
If logblock before LOG_ROLLUP (obsolated log), we don't call
replay_update_bitmap() for that logblock.

But we should call replay_update_bitmap() even if logblock is
obsolete. Otherwise, obsoleted logblock can be overwritten before next
rollup.

We are intentionally setting logcount which is including obsoleted
logblocks for debugging. And if we really want to remove obsoleted
logblocks, we should adjust logcount, not replay.

To fix, this move LOG_FREEBLOCKS to stage1, so stage2 is started with
freeblocks with latest rollup. And add all logblocks to derollup and
bitmap.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>