tux3: Refcount base delta
Now, while the backend is working, the frontend is blocked. Of course,
this stall makes worse latency and throughput.

To fix this stall, we have to use asynchronous task for backend. And
when we are using asynchronous backend, we have to keep delta boundary
(i.e. changes between change_{begin,end} must be in same delta).

To do it, we introduce reference count for delta number. Then, if
delta refcount become 0, we can start the backend for that delta.

Implementation detail is, delta is started with refcount == 1. And if
we decided to flush the delta, we decrement the initial refcount. So,
last referencer will see the refcount == 0, somewhen.

delta_get() - get reference count of current delta, then set grabbed
delta to current->journal.

delta_put() - put reference count of grabbed delta, then if refcount == 0,
set flag to tell pending backend job.

delta_transition() - prepare next delta, then switch current delta to
prepared delta. After that, decrement the initial refcount of old
current delta.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>