tux3: Reduce page->mapping, flush_list()
We are using the "page->mapping->host" to know the inode/mapping for
the buffer.

But, page->mapping is not reliable (e.g. truncate change it), we make
it stable though.

So, this series reduce the usage of "page->mapping" if possible.
This adds inode/mapping as parameter to pass those, and use those
instead of "page->mapping".

This adds "mapping" parameter to flush_list(), and reduce
"page->mapping" usage.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>