tux3: Don't pin log blocks, instead, use dirty state to pin
Now, log buffer was dirtied to write. So, we don't need to pin log
buffers explicitly. Instead, we have to clear dirty to discard log
buffers.

To discard log buffers, this adds new parameter for
log_finish_cycle(). If "discard" was specified, it clear dirty of log
buffers. So, as usual, clean log buffers are not written.

Also, as usual, dirty state prevents buffers to be reclaimed. So, we
don't need to pin log buffers anymore. But, replay path doesn't dirty
log buffers, so it has to pin buffer. Now, replay pins buffers
explicitly.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>