tux3: Fix inode->dirtied_when in the case of re-dirty
__mark_inode_dirty() doesn't know about delta boundary (we don't clear
I_DIRTY before flush, in order to prevent the inode to be freed). So,
if inode was re-dirtied for frontend delta while flushing old delta,
->dirtied_when may not be updated by __mark_inode_dirty() forever.

Although we don't use ->dirtied_when, bdi flusher uses ->dirtied_when
to decide flush timing, so we have to update ->dirtied_when ourself.

           cpu0                                      cpu1

__tux3_mark_inode_dirty()
    tux3_dirty_inode()
    __mark_inode_dirty()
    /* core dirty is still dirty */
                                      __tux3_clear_dirty_inode()
				      /* clear I_DIRTY and backend dirty */

So, if inode is not marked frontend dirty but was still marked as
I_DIRTY, this patch updates inode->dirtied_when (like __mark_inode_dirty()).

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>