tux3: Refactoring balloc stuff
Current balloc is not supporting multiple segments allocation at
once. So, this adds ability for it.

With this, we would be able to allocate more better block placement,
and even possibly have chance to optimize code itself.

Also, this separates balloc to 2 function:

    - find unused bits
    - modify bitmap

With this separation, we can allocate fewer than found unused bits.
This is needed for extent allocation when extent is not fit to leaf.

Signed-off-by: Daniel Phillips <daniel@tux3.org>
Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>