tux3: Try to create directories in less congested block groups
Try to create directories in less congested block groups. Choose a congestion
threshold based on directory depth (in root or not) and volume fullness. For
deeper directories and fuller volumes, accept fuller groups for the inode
number goal. Check a limited number of groups to find a good fit, otherwise
just use the extrapolated inum goal.

Signed-off-by: Daniel Phillips <d.phillips@partner.samsung.com>
Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>