tux3: Support mmap write: Add initial mmap write support
To support mmap write, we have to care about stabled pages (pages on
flushing delta). I.e. once pages are decided as flushing to backing
storage, we should not leak modification/truncation of pages in
frontend to flushing delta (this provides reliability like
data=journal).

To provide this, we do pagefork the page in previous delta to modify
in frontend.

Before issue I/O, PTE of the page become read-only. Then, we will
notice mmap write via page fault (->page_mkwrite() hook). In
->page_mkwrite(), if page was stabled (dirtied for previous delta, or
under I/O), we do pagefork the page to new page, switch PTE to point
new page, and make PTE writable.

With this, stabled page is not modified, and frontend can continue
modification to new page without waiting I/O.

[FIXME: ->page_mkwrite() can't support pagefork directly
(->page_mkwrite() is assuming the page in PTE is not modified.). So,
this patch switch PTE to new page, and make writable in
->page_mkwrite(). Then, retry page fault again, and resolve page fault
by switched PTE for new page.]

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>